pyproject.toml
[tool.poetry]
name = "autorag"
version = "0.1.0"
description = "Automatic hyperparameter optimization of your retrieval augmented generation pipeline."
authors = ["rasdani <73563550+rasdani@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
datasets = "^2.18.0"
python-dotenv = "^1.0.1"
matplotlib = "^3.8.3"
llama-index = "^0.10.19"
spacy = "^3.7.4"
transformers = "^4.38.2"
llama-index-embeddings-huggingface = "^0.1.4"
torch = "^2.2.1"
llama-index-llms-openrouter = "^0.1.4"
wandb = "^0.16.4"
llama-index-postprocessor-colbert-rerank = "^0.1.1"
fire = "^0.6.0"
sentence-transformers = "^2.5.1"
loguru = "^0.7.2"
einops = "^0.7.0"
llama-index-finetuning = "^0.1.4"
ruff = "^0.3.3"
llama-index-llms-litellm = "^0.1.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
